Pull up PruneColumns
